requirements.txt
torch
transformers
datasets
evaluate
opencv-python
fastapi
uvicorn
pillow
numpy
